refactor : update Contract Ui to depend on the interface insted of the implementation .
