+++ src/main/java/net/axel/services/implementations/ContractService.java
@@ -1,4 +1,4 @@
-package net.axel.services;
+package net.axel.services.implementations;
 
 import net.axel.models.dto.ContractDto;
 import net.axel.models.entities.Contract;
